+++ CHANGELOG.md
@@ -16,6 +16,7 @@ BUG FIXES:
   - When building toolchains with verbose mode, wait until output is fully
     read before moving on to next compilation.
   - Work with `-os` or `-arch` is an empty string.
+  - Building toolchain doesn't output "plan9" for all platforms.
 
 ## 0.1.0 (November 17, 2013)